rsync file to Iowa
git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@11835 e846bd91-c042-42a7-a218-4b65e2c1904d

+++ extensions/wikia/WikiFactory/Close/tarcopyimages.php
@@ -55,7 +55,7 @@ class CloseWikiTarAndCopyImages {
 				$dbname = $row->city_dbname;
 				$folder = WikiFactory::getVarValueByName( "wgUploadDirectory", $row->city_id );
 				if( $dbname && $folder ) {
-					Wikia::log( __METHOD__, "info", "city_id={$row->city_id} city_url={$row->city_url} city_dbname={$dbname} city_public={$row->city_public}");
+					Wikia::log( __CLASS__, "info", "city_id={$row->city_id} city_url={$row->city_url} city_dbname={$dbname} city_public={$row->city_public}");
 					$source = $this->tarFiles( $folder, $dbname );
 					$target = DumpsOnDemand::getUrl( $dbname, "images.tar", $this->mTarget );
 
@@ -70,8 +70,14 @@ class CloseWikiTarAndCopyImages {
 							$source,
 							escapeshellcmd( $target )
 						);
-						print $cmd  . "\n";
-						unlink( $source );
+						$output = wfShellExec( $cmd, $retval );
+						if( $retval > 0 ) {
+							Wikia::log( __CLASS__, "error", "{$cmd} command failed." );
+						}
+						else {
+							Wikia::log( __CLASS__, "info", "{$cource} copied to {$target}." );
+							unlink( $source );
+						}
 					}
 				}
 			}
